just publish everything :)
