+++ package.json
@@ -32,8 +32,7 @@
     "watch": "rollup -cw",
     "dist": "run-s lint types types:strict build build:types docs",
     "postdist": "copyfiles -f bundles/*/dist/* dist && copyfiles -f packages/*/dist/* dist/packages",
-    "codesandbox-ci": "npm run dist",
-    "postcodesandbox-ci": "ts-node ./scripts/codeSandboxCIRelease.ts",
+    "codesandbox-ci": "run-s dist prepublish-ci",
     "prerelease": "run-s clean:build test dist",
     "release": "ts-node ./scripts/release.ts",
     "prepublish-ci": "ts-node ./scripts/prepublish.ts",
