+++ scripts/prepublish.ts
@@ -1,6 +1,5 @@
 import path from 'path';
 import workspacesRun from 'workspaces-run';
-import { bumpDependencies } from './bumpDependencies';
 import { readJSON, writeJSON } from './utils/json';
 
 /** Before publish, let's replace all the relative file links with versions */
@@ -33,6 +32,23 @@ async function main()
 
         await writeJSON(path.join(workspace.dir, 'package.json'), workspace.config);
     });
+}
+
+/**
+ * Update the *dependencies in the package.json
+ * @param dependencies - Dependencies map
+ * @param version - Version to bump to
+ */
+export function bumpDependencies(dependencies: Record<string, string> = {}, version: string)
+{
+    Object.keys(dependencies)
+        // Only consider bumping monorepo packages
+        .filter((n) => dependencies[n].startsWith('file:'))
+        .forEach((n) =>
+        {
+            // replace with exact version
+            dependencies[n] = version;
+        });
 }
 
 main();
